requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.2
botocore==1.39.2
click==8.2.1
confluent-kafka==2.11.0
dnspython==2.7.0
fastapi==0.115.14
h11==0.16.0
idna==3.10
jmespath==1.0.1
pika==1.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
redis==6.2.0
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tinydb==4.8.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
